Change ClosenessVertex to work like IntIntNullVertex
Also switch to OpenIntIntHashMap for the shortest paths and to integers
for all the integral vertex values, should be enough and uses half the
space.

Also change BasicVertex and descendants because putMessages()
and releaseResources() were package private, this intefered with
implementing the custom BasicVertex subclass outside of the "graph"
package.