Fix bug with missing entries in shortes paths map
This bug that was discovered by Zachary Terner was due to the fact that
the compute function would only run when a vertex had new messages. This
is still the case but we now update the shortest paths map to insert all
the missing values when we add a new value in an iteration.